refactor(core): split network types.rs into a types/ directory module
