+++ crates/rustnet-core/src/network/types/mod.rs
@@ -1,0 +1,17 @@
+//! Shared connection and protocol types: connection identity and state,
+//! per-protocol DPI info, traffic history, rate tracking, and RTT
+//! measurement.
+
+mod connection;
+mod graph;
+mod identity;
+mod protocol_info;
+mod rates;
+mod rtt;
+
+pub use connection::*;
+pub use graph::*;
+pub use identity::*;
+pub use protocol_info::*;
+pub use rates::*;
+pub use rtt::*;
